Connect PsyAudioDevice with the PsyAudioMixer

+++ psy/psy-queue.h
@@ -13,7 +13,7 @@ G_BEGIN_DECLS
 typedef struct PsyAudioQueue PsyAudioQueue;
 
 G_MODULE_EXPORT PsyAudioQueue *
-psy_audio_queue_new(gint num_samples);
+psy_audio_queue_new(gsize num_samples);
 
 G_MODULE_EXPORT void
 psy_audio_queue_free(PsyAudioQueue *self);
